Move functionality to a modifer, have component version use the modifier
