+++ .gitignore
@@ -23,3 +23,5 @@
 /.node_modules.ember-try/
 /bower.json.ember-try
 /package.json.ember-try
+
+/.idea/
