+++ addon/components/litepicker.hbs
@@ -1,7 +1,5 @@
 <Input
   class="ember-litepicker-input"
   ...attributes
-  {{did-insert this._initializeOptions}}
-  {{did-update this._updateOptions @format @lang @startDate @endDate}}
-  {{will-destroy this._destroyOptions}}
+  {{litepicker this.args}}
 />
