+++ app/modifiers/litepicker.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-litepicker/modifiers/litepicker';
