+++ package.json
@@ -32,6 +32,7 @@
     "@ember/render-modifiers": "^1.0.2",
     "ember-cli-babel": "^7.21.0",
     "ember-cli-htmlbars": "^5.1.2",
+    "ember-modifier": "^1.0.3",
     "litepicker": "^1.5.5"
   },
   "devDependencies": {
