+++ tests/dummy/app/templates/application.hbs
@@ -1,1 +1,7 @@
-<Litepicker @singleMode={{false}} @startDate="23.12.2019" @endDate="28.12.2019" autocomplete="off" />
+<div>
+  Litepicker as a component <Litepicker @singleMode={{false}} @startDate="23.12.2019" @endDate="28.12.2019" autocomplete="off" />
+</div>
+<div>&nbsp;</div>
+<div>
+  Litepicker as a modifier <Input {{litepicker singleMode=false startDate="23.12.2019" endDate="28.12.2019" autocomplete="off"}} />
+</div>
